Reattempt downloads in batches
Previously downloading would count as failed as long as a single one
failed. This would mean that already succeeded downloads would be
reattempted on the next go. yt-dlp sometimes catches these but the next
go could either be on another day or the file format and extension could
be different, leading to duplicate downloads again.

This is somewhat mitigated by downloading in smaller batches, meaning a
failed download only invalidates that batch.