Cache and display channel and video names
Previously the list of channels at the top of the main interface as well
as the output when reattempting was quite meaningless as it only
displayed ids. Since querying the information each time it is displayed
would involve network requests and would therefore lead to serious
delays, a far more reasonable thing to do is to cache the info when it
is available anyway, e.g. when watching for changes, and just displaying
those cached values later. Since a channel name or video title does not
change that often those cached values should be very accurate anyways.